Added auth

+++ src/roles/roles.service.ts
@@ -1,7 +1,10 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { classToPlain, plainToClass } from 'class-transformer';
+import { ResourceNotFoundError } from '../resource-not-found.error';
 import { Repository } from 'typeorm';
 import { CreateRoleDto } from './dto/create-role.dto';
+import { RoleDto } from './dto/role.dto';
 import { Role } from './role.entity';
 
 @Injectable()
@@ -11,24 +14,37 @@ export class RolesService {
     private rolesRepository: Repository<Role>,
   ) {}
 
-  async findAll(): Promise<Role[]> {
-    return this.rolesRepository.find();
+  async findAll(): Promise<RoleDto[]> {
+    const roles = await this.rolesRepository.find();
+    return roles.map(this.convertToDto);
   }
 
-  async findOne(id: string): Promise<Role> {
-    return this.rolesRepository.findOneOrFail(id);
+  async findOne(id: string): Promise<RoleDto> {
+    const role = await this.rolesRepository.findOne(id);
+    if (!role) throw new ResourceNotFoundError();
+
+    return role;
   }
 
-  async create(role: CreateRoleDto): Promise<Role> {    
+  async create(role: CreateRoleDto): Promise<RoleDto> {
     return this.rolesRepository.save(role);
   }
 
-  async update(id: string, role: CreateRoleDto): Promise<Role> {    
-    return this.rolesRepository.save({ ...role, id: Number(id) });
+  async update(id: string, roleDto: RoleDto): Promise<RoleDto> {
+    const role = await this.rolesRepository.findOne(id);
+    if (!role) throw new ResourceNotFoundError();
+
+    return this.rolesRepository.save({ ...roleDto, id: Number(id) });
   }
 
   async remove(id: string): Promise<void> {
-    await this.rolesRepository.findOneOrFail(id);
+    const role = await this.rolesRepository.findOne(id);
+    if (!role) throw new ResourceNotFoundError();
+
     await this.rolesRepository.delete(id);
+  }
+
+  convertToDto(role: any): RoleDto {
+    return plainToClass(RoleDto, classToPlain(role));
   }
 }
